+++ tests/vibe3/orchestra/test_dispatch_queue_operations.py
@@ -2,18 +2,11 @@
 
 from __future__ import annotations
 
-from unittest.mock import MagicMock
-
 import pytest
 
 from vibe3.models.orchestration import IssueState
 from vibe3.orchestra.global_dispatch_coordinator import (
     QueueEntry,
-)
-from vibe3.orchestra.queue_operations import (
-    _auto_resume_to_ready,
-    _last_auto_resume_attempt,
-    select_ready_issues_from_collected_issues,
 )
 
 
@@ -358,633 +351,3 @@ class TestQueueOperations:
         await coordinator.coordinate()
 
         assert len(emit_calls) == 0
-
-
-class TestAutoResumeNoFlowScene:
-    """Tests for auto-resume of orphaned issues without flow scene."""
-
-    def test_auto_resume_claimed_no_branch(self, make_issue_info, monkeypatch) -> None:
-        """Issue with CLAIMED state and no branch triggers auto-resume."""
-        from vibe3.models.orchestra_config import OrchestraConfig
-
-        config = OrchestraConfig(repo="owner/repo")
-        issue = make_issue_info(100, IssueState.CLAIMED)
-
-        mock_label_service = MagicMock()
-        mock_label_service.transition = MagicMock()
-
-        def mock_label_service_init(*args, **kwargs):
-            return mock_label_service
-
-        monkeypatch.setattr(
-            "vibe3.services.label_service.LabelService",
-            mock_label_service_init,
-        )
-
-        event_calls = []
-
-        def mock_append_event(source: str, message: str) -> None:
-            event_calls.append((source, message))
-
-        monkeypatch.setattr(
-            "vibe3.orchestra.queue_operations.append_orchestra_event",
-            mock_append_event,
-        )
-
-        _auto_resume_to_ready(issue, config)
-
-        mock_label_service.transition.assert_called_once_with(
-            100,
-            IssueState.READY,
-            actor="orchestra:auto-resume",
-            force=True,
-        )
-
-        assert len(event_calls) == 1
-        assert event_calls[0][0] == "dispatcher"
-        assert "auto-resume #100" in event_calls[0][1]
-        assert "state=claimed" in event_calls[0][1]
-        assert "recovered to ready" in event_calls[0][1]
-
-    def test_no_auto_resume_ready_no_branch(self, make_issue_info, monkeypatch) -> None:
-        """Issue with READY state and no branch does NOT trigger auto-resume."""
-        from vibe3.models.orchestra_config import OrchestraConfig
-
-        config = OrchestraConfig(repo="owner/repo")
-        issue = make_issue_info(101, IssueState.READY)
-
-        mock_label_service = MagicMock()
-        mock_label_service.transition = MagicMock()
-
-        def mock_label_service_init(*args, **kwargs):
-            return mock_label_service
-
-        monkeypatch.setattr(
-            "vibe3.services.label_service.LabelService",
-            mock_label_service_init,
-        )
-
-        # Call the function with READY issue
-        _auto_resume_to_ready(issue, config)
-
-        # Verify transition is NOT called (defensive guard)
-        mock_label_service.transition.assert_not_called()
-
-    def test_no_auto_resume_blocked_no_branch(
-        self, make_issue_info, monkeypatch
-    ) -> None:
-        """Issue with BLOCKED state and no branch does NOT trigger auto-resume."""
-        from vibe3.models.orchestra_config import OrchestraConfig
-
-        config = OrchestraConfig(repo="owner/repo")
-        issue = make_issue_info(102, IssueState.BLOCKED)
-
-        mock_label_service = MagicMock()
-        mock_label_service.transition = MagicMock()
-
-        def mock_label_service_init(*args, **kwargs):
-            return mock_label_service
-
-        monkeypatch.setattr(
-            "vibe3.services.label_service.LabelService",
-            mock_label_service_init,
-        )
-
-        # Call the function with BLOCKED issue
-        _auto_resume_to_ready(issue, config)
-
-        # Verify transition is NOT called (state machine invariant)
-        mock_label_service.transition.assert_not_called()
-
-    def test_auto_resume_failure_does_not_crash(
-        self, make_issue_info, monkeypatch
-    ) -> None:
-        """LabelService.transition() raises exception — verify log and no crash."""
-        from vibe3.models.orchestra_config import OrchestraConfig
-
-        config = OrchestraConfig(repo="owner/repo")
-        issue = make_issue_info(103, IssueState.CLAIMED)
-
-        mock_label_service = MagicMock()
-        mock_label_service.transition = MagicMock(side_effect=RuntimeError("API error"))
-
-        def mock_label_service_init(*args, **kwargs):
-            return mock_label_service
-
-        monkeypatch.setattr(
-            "vibe3.services.label_service.LabelService",
-            mock_label_service_init,
-        )
-
-        event_calls = []
-
-        def mock_append_event(source: str, message: str) -> None:
-            event_calls.append((source, message))
-
-        monkeypatch.setattr(
-            "vibe3.orchestra.queue_operations.append_orchestra_event",
-            mock_append_event,
-        )
-
-        # Should not raise exception
-        _auto_resume_to_ready(issue, config)
-
-        mock_label_service.transition.assert_called_once()
-
-        assert len(event_calls) == 1
-        assert event_calls[0][0] == "dispatcher"
-        assert "auto-resume #103 failed" in event_calls[0][1]
-        assert "API error" in event_calls[0][1]
-
-    def test_auto_resume_skipped_for_manager_role(
-        self, make_issue, make_issue_info, make_capacity, make_coordinator, monkeypatch
-    ) -> None:
-        """Manager role dispatch skips branch check entirely, no auto-resume."""
-        coordinator = make_coordinator(
-            "manager",
-            capacity=make_capacity(remaining=1),
-            mock_health_check=True,
-        )
-
-        # Manager role should not check branches at all
-        issues = [
-            make_issue_info(200, IssueState.READY, assignees=[]),
-        ]
-
-        mock_label_service = MagicMock()
-        mock_label_service.transition = MagicMock()
-
-        def mock_label_service_init(*args, **kwargs):
-            return mock_label_service
-
-        monkeypatch.setattr(
-            "vibe3.services.label_service.LabelService",
-            mock_label_service_init,
-        )
-
-        event_calls = []
-
-        def mock_append_event(source: str, message: str) -> None:
-            event_calls.append((source, message))
-
-        monkeypatch.setattr(
-            "vibe3.orchestra.queue_operations.append_orchestra_event",
-            mock_append_event,
-        )
-
-        # Call the queue selector directly
-        select_ready_issues_from_collected_issues(
-            issues=issues,
-            trigger_state=IssueState.READY,
-            config=coordinator._config,
-            github=coordinator._github,
-            store=coordinator._store,
-            flow_manager=coordinator._flow_manager,
-            qualify_gate=coordinator._qualify_gate,
-            supervisor_label=coordinator._config.supervisor_handoff.issue_label,
-        )
-
-        # Manager role skips branch check, so no auto-resume should be triggered
-        mock_label_service.transition.assert_not_called()
-
-        # No auto-resume events should be logged
-        auto_resume_calls = [call for call in event_calls if "auto-resume" in call[1]]
-        assert len(auto_resume_calls) == 0
-
-    def test_auto_resume_triggered_for_claimed_no_branch_integration(
-        self, make_issue_info, make_capacity, make_coordinator, monkeypatch
-    ) -> None:
-        """Integration test: CLAIMED issue with no branch triggers auto-resume."""
-        coordinator = make_coordinator(
-            "planner",
-            capacity=make_capacity(remaining=1),
-            mock_health_check=True,
-        )
-
-        # Setup CLAIMED issue with no branch (orphaned)
-        issues = [
-            make_issue_info(300, IssueState.CLAIMED),
-        ]
-
-        # Mock get_flow_context to return no branch
-        def mock_get_flow_context(*args, **kwargs):
-            return (None, None)
-
-        monkeypatch.setattr(
-            "vibe3.orchestra.queue_operations.get_flow_context",
-            mock_get_flow_context,
-        )
-
-        # Mock qualify gate to return the trigger state (allow issue through)
-        coordinator._qualify_gate.run_qualify_gate = MagicMock(
-            return_value=IssueState.CLAIMED
-        )
-
-        mock_label_service = MagicMock()
-        mock_label_service.transition = MagicMock()
-
-        def mock_label_service_init(*args, **kwargs):
-            return mock_label_service
-
-        monkeypatch.setattr(
-            "vibe3.services.label_service.LabelService",
-            mock_label_service_init,
-        )
-
-        event_calls = []
-
-        def mock_append_event(source: str, message: str) -> None:
-            event_calls.append((source, message))
-
-        monkeypatch.setattr(
-            "vibe3.orchestra.queue_operations.append_orchestra_event",
-            mock_append_event,
-        )
-
-        # Call the queue selector - should trigger auto-resume
-        select_ready_issues_from_collected_issues(
-            issues=issues,
-            trigger_state=IssueState.CLAIMED,
-            config=coordinator._config,
-            github=coordinator._github,
-            store=coordinator._store,
-            flow_manager=coordinator._flow_manager,
-            qualify_gate=coordinator._qualify_gate,
-            supervisor_label=coordinator._config.supervisor_handoff.issue_label,
-        )
-
-        # Verify auto-resume was triggered
-        mock_label_service.transition.assert_called_once_with(
-            300,
-            IssueState.READY,
-            actor="orchestra:auto-resume",
-            force=True,
-        )
-
-        # Verify event logged
-        auto_resume_calls = [
-            call for call in event_calls if "auto-resume #300" in call[1]
-        ]
-        assert len(auto_resume_calls) == 1
-
-    def test_auto_resume_not_triggered_for_blocked_no_branch_integration(
-        self, make_issue_info, make_capacity, make_coordinator, monkeypatch
-    ) -> None:
-        """Integration test: BLOCKED issue with no branch does NOT trigger."""
-        coordinator = make_coordinator(
-            "planner",
-            capacity=make_capacity(remaining=1),
-            mock_health_check=True,
-        )
-
-        # Setup BLOCKED issue with no branch
-        issues = [
-            make_issue_info(301, IssueState.BLOCKED),
-        ]
-
-        # Mock get_flow_context to return no branch
-        def mock_get_flow_context(*args, **kwargs):
-            return (None, None)
-
-        monkeypatch.setattr(
-            "vibe3.orchestra.queue_operations.get_flow_context",
-            mock_get_flow_context,
-        )
-
-        # Mock qualify gate to return the trigger state (allow issue through)
-        coordinator._qualify_gate.run_qualify_gate = MagicMock(
-            return_value=IssueState.BLOCKED
-        )
-
-        mock_label_service = MagicMock()
-        mock_label_service.transition = MagicMock()
-
-        def mock_label_service_init(*args, **kwargs):
-            return mock_label_service
-
-        monkeypatch.setattr(
-            "vibe3.services.label_service.LabelService",
-            mock_label_service_init,
-        )
-
-        event_calls = []
-
-        def mock_append_event(source: str, message: str) -> None:
-            event_calls.append((source, message))
-
-        monkeypatch.setattr(
-            "vibe3.orchestra.queue_operations.append_orchestra_event",
-            mock_append_event,
-        )
-
-        # Call the queue selector - should NOT trigger auto-resume
-        select_ready_issues_from_collected_issues(
-            issues=issues,
-            trigger_state=IssueState.BLOCKED,
-            config=coordinator._config,
-            github=coordinator._github,
-            store=coordinator._store,
-            flow_manager=coordinator._flow_manager,
-            qualify_gate=coordinator._qualify_gate,
-            supervisor_label=coordinator._config.supervisor_handoff.issue_label,
-        )
-
-        # Verify auto-resume was NOT triggered (state machine invariant)
-        mock_label_service.transition.assert_not_called()
-
-        # Verify no auto-resume events logged
-        auto_resume_calls = [call for call in event_calls if "auto-resume" in call[1]]
-        assert len(auto_resume_calls) == 0
-
-
-class TestAutoResumeCooldown:
-    """Tests for auto-resume cooldown mechanism."""
-
-    def test_cooldown_skips_rapid_retry(self, make_issue_info, monkeypatch) -> None:
-        """Second auto-resume within cooldown period is skipped (failure case)."""
-        from vibe3.models.orchestra_config import OrchestraConfig
-
-        config = OrchestraConfig(repo="owner/repo")
-        issue = make_issue_info(100, IssueState.CLAIMED)
-
-        # Start with empty cooldown dict for test isolation
-        monkeypatch.setattr(
-            "vibe3.orchestra.queue_operations._last_auto_resume_attempt",
-            {},
-        )
-
-        mock_label_service = MagicMock()
-        # First call fails, second would succeed
-        mock_label_service.transition = MagicMock(
-            side_effect=[RuntimeError("API error"), None]
-        )
-
-        def mock_label_service_init(*args, **kwargs):
-            return mock_label_service
-
-        monkeypatch.setattr(
-            "vibe3.services.label_service.LabelService",
-            mock_label_service_init,
-        )
-
-        event_calls = []
-
-        def mock_append_event(source: str, message: str) -> None:
-            event_calls.append((source, message))
-
-        monkeypatch.setattr(
-            "vibe3.orchestra.queue_operations.append_orchestra_event",
-            mock_append_event,
-        )
-
-        # Mock time to control cooldown
-        current_time = [1000.0]
-
-        def mock_time():
-            return current_time[0]
-
-        monkeypatch.setattr(
-            "vibe3.orchestra.queue_operations.time.time",
-            mock_time,
-        )
-
-        # First call - fails but sets cooldown
-        _auto_resume_to_ready(issue, config)
-
-        # Advance time by 60s (within cooldown)
-        current_time[0] = 1060.0
-
-        # Second call - should be skipped by cooldown (transition not called)
-        _auto_resume_to_ready(issue, config)
-
-        # Verify transition was called only once (second was skipped by cooldown)
-        mock_label_service.transition.assert_called_once_with(
-            100,
-            IssueState.READY,
-            actor="orchestra:auto-resume",
-            force=True,
-        )
-
-    def test_cooldown_allows_after_expiry(self, make_issue_info, monkeypatch) -> None:
-        """Auto-resume after cooldown expiry is allowed."""
-        from vibe3.models.orchestra_config import OrchestraConfig
-
-        config = OrchestraConfig(repo="owner/repo")
-        issue = make_issue_info(101, IssueState.CLAIMED)
-
-        # Start with empty cooldown dict for test isolation
-        monkeypatch.setattr(
-            "vibe3.orchestra.queue_operations._last_auto_resume_attempt",
-            {},
-        )
-
-        mock_label_service = MagicMock()
-        mock_label_service.transition = MagicMock()
-
-        def mock_label_service_init(*args, **kwargs):
-            return mock_label_service
-
-        monkeypatch.setattr(
-            "vibe3.services.label_service.LabelService",
-            mock_label_service_init,
-        )
-
-        event_calls = []
-
-        def mock_append_event(source: str, message: str) -> None:
-            event_calls.append((source, message))
-
-        monkeypatch.setattr(
-            "vibe3.orchestra.queue_operations.append_orchestra_event",
-            mock_append_event,
-        )
-
-        # Mock time to control cooldown
-        current_time = [1000.0]
-
-        def mock_time():
-            return current_time[0]
-
-        monkeypatch.setattr(
-            "vibe3.orchestra.queue_operations.time.time",
-            mock_time,
-        )
-
-        # First call
-        _auto_resume_to_ready(issue, config)
-
-        # Advance time by 301s (beyond cooldown)
-        current_time[0] = 1301.0
-
-        # Second call - should be allowed
-        _auto_resume_to_ready(issue, config)
-
-        # Verify transition was called twice
-        assert mock_label_service.transition.call_count == 2
-
-    def test_successful_resume_clears_cooldown(
-        self, make_issue_info, monkeypatch
-    ) -> None:
-        """Successful resume clears cooldown for immediate future resume."""
-        from vibe3.models.orchestra_config import OrchestraConfig
-
-        config = OrchestraConfig(repo="owner/repo")
-        issue = make_issue_info(102, IssueState.CLAIMED)
-
-        # Start with empty cooldown dict for test isolation
-        monkeypatch.setattr(
-            "vibe3.orchestra.queue_operations._last_auto_resume_attempt",
-            {},
-        )
-
-        mock_label_service = MagicMock()
-        mock_label_service.transition = MagicMock()
-
-        def mock_label_service_init(*args, **kwargs):
-            return mock_label_service
-
-        monkeypatch.setattr(
-            "vibe3.services.label_service.LabelService",
-            mock_label_service_init,
-        )
-
-        event_calls = []
-
-        def mock_append_event(source: str, message: str) -> None:
-            event_calls.append((source, message))
-
-        monkeypatch.setattr(
-            "vibe3.orchestra.queue_operations.append_orchestra_event",
-            mock_append_event,
-        )
-
-        # Mock time
-        current_time = [1000.0]
-
-        def mock_time():
-            return current_time[0]
-
-        monkeypatch.setattr(
-            "vibe3.orchestra.queue_operations.time.time",
-            mock_time,
-        )
-
-        # First call - succeeds and clears cooldown
-        _auto_resume_to_ready(issue, config)
-
-        # No time advance - should still work because cooldown was cleared
-        _auto_resume_to_ready(issue, config)
-
-        # Verify transition was called twice (cooldown cleared after success)
-        assert mock_label_service.transition.call_count == 2
-
-    def test_cooldown_tracks_per_issue_independently(
-        self, make_issue_info, monkeypatch
-    ) -> None:
-        """Cooldown is tracked independently per issue."""
-        from vibe3.models.orchestra_config import OrchestraConfig
-
-        config = OrchestraConfig(repo="owner/repo")
-        issue_100 = make_issue_info(100, IssueState.CLAIMED)
-        issue_200 = make_issue_info(200, IssueState.CLAIMED)
-
-        # Start with empty cooldown dict for test isolation
-        monkeypatch.setattr(
-            "vibe3.orchestra.queue_operations._last_auto_resume_attempt",
-            {},
-        )
-
-        mock_label_service = MagicMock()
-        mock_label_service.transition = MagicMock()
-
-        def mock_label_service_init(*args, **kwargs):
-            return mock_label_service
-
-        monkeypatch.setattr(
-            "vibe3.services.label_service.LabelService",
-            mock_label_service_init,
-        )
-
-        event_calls = []
-
-        def mock_append_event(source: str, message: str) -> None:
-            event_calls.append((source, message))
-
-        monkeypatch.setattr(
-            "vibe3.orchestra.queue_operations.append_orchestra_event",
-            mock_append_event,
-        )
-
-        # Mock time
-        current_time = [1000.0]
-
-        def mock_time():
-            return current_time[0]
-
-        monkeypatch.setattr(
-            "vibe3.orchestra.queue_operations.time.time",
-            mock_time,
-        )
-
-        # Call for issue #100
-        _auto_resume_to_ready(issue_100, config)
-
-        # Immediately call for issue #200 - should work (different issue)
-        _auto_resume_to_ready(issue_200, config)
-
-        # Verify both calls went through
-        assert mock_label_service.transition.call_count == 2
-
-    def test_cooldown_evicts_stale_entries(self, make_issue_info, monkeypatch) -> None:
-        """Stale cooldown entries (>24h) are evicted to bound memory."""
-        from vibe3.models.orchestra_config import OrchestraConfig
-
-        config = OrchestraConfig(repo="owner/repo")
-        issue = make_issue_info(300, IssueState.CLAIMED)
-
-        # Seed stale entry from 25 hours ago
-        stale_time = 1000.0
-        monkeypatch.setattr(
-            "vibe3.orchestra.queue_operations._last_auto_resume_attempt",
-            {300: stale_time, 999: stale_time},
-        )
-
-        mock_label_service = MagicMock()
-        mock_label_service.transition = MagicMock()
-
-        def mock_label_service_init(*args, **kwargs):
-            return mock_label_service
-
-        monkeypatch.setattr(
-            "vibe3.services.label_service.LabelService",
-            mock_label_service_init,
-        )
-
-        event_calls = []
-
-        def mock_append_event(source: str, message: str) -> None:
-            event_calls.append((source, message))
-
-        monkeypatch.setattr(
-            "vibe3.orchestra.queue_operations.append_orchestra_event",
-            mock_append_event,
-        )
-
-        # Current time is 25h after stale entries
-        current_time = [stale_time + 86400 + 1]
-
-        def mock_time():
-            return current_time[0]
-
-        monkeypatch.setattr(
-            "vibe3.orchestra.queue_operations.time.time",
-            mock_time,
-        )
-
-        # Call triggers eviction + proceeds with resume
-        _auto_resume_to_ready(issue, config)
-
-        # Entry for 999 should be evicted (stale)
-        assert 999 not in _last_auto_resume_attempt
-        # Issue 300 was resumed successfully (pop clears on success)
-        mock_label_service.transition.assert_called_once()
